Move conventions to Tracer and support settings
