--- src/Datadog.Trace.ClrProfiler.Managed/Conventions/IOutboundHttpConvention.cs
@@ -1,7 +1,0 @@
-namespace Datadog.Trace.ClrProfiler.Conventions
-{
-    internal interface IOutboundHttpConvention
-    {
-        void Apply(OutboundHttpArgs args);
-    }
-}
